refactor: ECSの初期化から、権限部分だけを分離する。
